Add PriceVolatilityOracle
